rename console-commander to consoleCommander
