+++ src/core/shared/utility/consoleCommander.ts
